+++ CHANGES.rst
@@ -25,3 +25,17 @@
 
 - RelStorage: The value of ``sql_adapter_extra_args`` is validated to
   be syntactically correct.
+
+- RelStorage: Support providing a PostgreSQL DSN
+
+- All storages: Change the default to use zlibstorage only to
+  decompress existing records, not to compress new records. Set the
+  ``compress`` option (in this recipe or the ``environment`` recipe)
+  to ``compress`` to turn compression on. Set it to ``decompress`` to
+  explicitly request only decompression, and set it to ``none`` to
+  explicitly disable all compression and decompression. In the future,
+  expect the default to change to ``none``. See `issue 9 <https://github.com/NextThought/nti.recipes.zodb/issues/9>`_.
+
+- ZEO: Instead of using ``${buildout:directory}/var/`` and
+  ``${buildout:directory}/var/log`` directly, refer
+  to ``${deployment:run-directory}`` and ``${deployment:log-directory}``.